change in the interface of operator[]
git-svn-id: https://svn.salilab.org/imp/trunk@2510 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ embed/trunk/src/emlib/Vector3.h
@@ -182,7 +182,7 @@ public:
 
   //// Treats vector like a real array. Acceptable indices 0, 1, 2. Other
   // indices may cause run-time errors. 
-  const real& operator[](const unsigned short coord) const{
+  const real& operator[](const unsigned int coord) const{
     return c[coord];
   }
